requirements.txt
streamlit
openai
qdrant-client
fastembed
python-dotenv
